requirements.txt
python-dotenv~=0.15.0
cloudinary~=1.24.0
requests~=2.25.1
Pillow~=8.0.1